==> caQTL_mapping_pbenaglio/notebooks/0_Input_SNPs_WASP.ipynb <==
library(parallel)
library(stringr)

vcf_dir = '/nfs/lab/projects/pbmc_snATAC/pipeline/snp_array/final_vcfs/'

outdir = '/nfs/lab/projects/pbmc_snATAC/analysis_v2/wasp_peaks/filtered_vcfs'

bed     =  '/nfs/lab/projects/pbmc_snATAC/analysis_v2/peaks/pbmc.sorted.100pad.merged.bed'

### FIlter original vcf files for just  snps in peaks, imputation quality and heterozygous genotype and remove indels

    for (n in 1:22) {

    vcf    = paste0(vcf_dir, "/chr",n ,'.PBMC.topmed.hg19.final.vcf.gz')
      
    outvcf = paste0(outdir, "/chr",n ,'.filt.vcf.gz')
    system(paste('bcftools view', vcf ,
    '-R' , bed, 
    "-i \'INFO/R2[0]>0.7\' | bcftools view --types snps -g het -Oz -o" , outvcf))
 
        system(paste("tabix", outvcf))
    }

#### actually keep the indels
 for (n in 1:22) {

    vcf    = paste0(vcf_dir, "/chr",n ,'.PBMC.topmed.hg19.final.vcf.gz')
      
    outvcf = paste0(outdir, "/chr",n ,'.filt.vcf.gz')
    system(paste('bcftools view', vcf ,
    '-R' , bed, 
    "-i \'INFO/R2[0]>0.7\' | bcftools view -g het -Oz -o" , outvcf))
 
        system(paste("tabix", outvcf))
    }


ids = c(1,3:6, 8:10,12:15)

for (n in ids) {
    from   = paste0('/nfs/lab/projects/pbmc_snATAC/analysis_v2/wasp_peaks/PBMC', n, '/sc_bams')
    newdir = paste0('/nfs/lab/projects/pbmc_snATAC/analysis_v2/wasp_peaks/singlecell/PBMC', n)
    system(paste("mkdir", newdir))
   
    system(paste("mv", from, newdir))
    
    }

dir = '/nfs/lab/projects/pbmc_snATAC/analysis_v2/wasp_peaks/bulk/PBMC3'

tab = data.frame()
for(n in 1:22){
tab1 = read.table(paste0(dir, "/snps/chr", n, ".snps.txt.gz"))
tab1$chr = n    
tab = rbind(tab, tab1)
}

dim(tab)

tab$size = apply(tab, 1, function(x) max(nchar(x[2]), nchar(x[3])))

varlength = table(tab$size)

varlength

1-sum((varlength[1])/sum(varlength))

sum(varlength[2:(length(varlength))])

tab$SNP = paste(tab$chr, tab$V1,tab$V2,tab$V3, sep=":")

head(tab)

bed = data.frame(tab$chr, tab$V1-1, tab$V1, tab$SNP)

bed =subset(bed, tab$size>1)

dim(bed)

write.table(bed, '/nfs/lab/projects/pbmc_snATAC/analysis_v2/example_boxplots/indels.bed', sep="\t", quote=F, 
            col.names=F, row.names=F)

examp = read.table('/nfs/lab/projects/pbmc_snATAC/analysis_v2/example_boxplots/snp_list')

ex_bed = data.frame(examp[,1], examp[,2]-10000, examp[,2]+10000, paste(examp[,1], examp[,2], sep=":"))

write.table(ex_bed, '/nfs/lab/projects/pbmc_snATAC/analysis_v2/example_boxplots/snp_list_10kb.bed', sep="\t", quote=F, 
            col.names=F, row.names=F)

bedtools intersect -a snp_list_10kb.bed -b indels.bed -wa -wo > indels_at_snps.bed
