==> caQTL_mapping_pbenaglio/notebooks/0_wasp_prepare.ipynb <==
library(stringr)

setwd('/nfs/lab/projects/pbmc_snATAC/analysis_v2/wasp')

bcs = read.table("pbmc1-15.clustering.FinalGood.Barcodes.txt", header=T)

bcs$index = toupper(bcs$sample)

bcs$donor = str_split_fixed(bcs$index, "_", 2)[,1]

table(bcs$donor)

write.table(bcs[,2:3], "PBMC.barcode.list.txt", sep="\t", quote=F, row.names=F)

x=15
system(paste0("ls -lh PBMC",x, "/remap > pbmc",x,".txt"))
pb1 = read.table(paste0("pbmc",x,".txt"), skip=1)
pb1$bc = str_split_fixed(pb1$V9, "\\.",2)[,1]
bc1 = bcs$index[bcs$donor==paste0("PBMC",x)]

sum(!duplicated(bc1))

sum(!duplicated(pb1$bc))

head(pb1)

old = read.table('/nfs/lab/projects/pbmc_snATAC/pipeline/snATAC/combined_files/pbmc1to12/barcode_and_cluster_label_files/main_cluster_label_files/pbmc1to12.cluster_labels_v6_afterDoublet9_microGroups.txt', 
                          sep="\t",header=T)

bar = str_split_fixed(old$index, "\\_",2)
old$sample  = bar[,1]

table(old$sample)
