==> caQTL_mapping_pbenaglio/notebooks/3_EUR_Prepare_vcf.ipynb <==
Requires that the ASE was calculated on all samples in the other notebook

library(parallel)
library(stringr)

homedir = '/nfs/lab/projects/pbmc_snATAC/analysis_v2/rasqual_eur/'

alldir = '/nfs/lab/projects/pbmc_snATAC/analysis_v2/rasqual/'

setwd(homedir)

filter_ase_vcf = function (reso = "fine"){
   
    celltypes = list.files(reso)[!grepl( ".txt",list.files(reso))]
    
    for (cell in celltypes){
        cat(cell,"\n")
        outdir  = paste(homedir, reso, cell, "vcfs_peaksonly", sep="/")
        invcf   = paste(alldir, reso, cell,"vcfs_peaksonly/full.ase.filtered.vcf.gz"  ,sep="/")
        outvcf  = paste(homedir, reso, cell,"vcfs_peaksonly/full.ase.filtered.vcf.gz"  ,sep="/")
        bed     = paste0(homedir, reso, "/",cell, "/" , cell, '.filtered.bed') 
         
        system(paste("mkdir -p", outdir))
        
        system(paste('bcftools view', invcf ,
        '-R' , bed, 
        '-s PBMC2,PBMC3,PBMC4,PBMC5,PBMC8,PBMC9,PBMC10,PBMC12,PBMC13,PBMC14',
        "| /home/pbenaglio/bin/bcftools view -i \'COUNT(GT=\"het\")>1\' -Oz -o", outvcf))
        ### note a newer version (1.9) of bcftools was used to use the COUNT command.
        
        system(paste("tabix", outvcf, "-f"))
        
    }
    
}

filter_ase_vcf(reso="fine")

filter_ase_vcf(reso="broad")

filter_ase_vcf(reso="bulk")

   celltypes = "t"
   reso      = "broad"
    
    for (cell in celltypes){
        cat(cell,"\n")
        outdir  = paste(homedir, reso, cell, "vcfs_peaksonly", sep="/")
        invcf   = paste(alldir, reso, cell,"vcfs_peaksonly/full.ase.filtered.vcf.gz"  ,sep="/")
        outvcf  = paste(homedir, reso, cell,"vcfs_peaksonly/full.ase.filtered.vcf.gz"  ,sep="/")
        bed     = paste0(homedir, reso, "/",cell, "/" , cell, '.filtered.bed') 
         
        system(paste("mkdir -p", outdir))
        
        system(paste('bcftools view', invcf ,
        '-R' , bed, 
        '-s PBMC2,PBMC3,PBMC4,PBMC5,PBMC8,PBMC9,PBMC10,PBMC12,PBMC13,PBMC14',
        "| /home/pbenaglio/bin/bcftools view -i \'COUNT(GT=\"het\")>1\' -Oz -o", outvcf))
        ### note a newer version (1.9) of bcftools was used to use the COUNT command.
        
        system(paste("tabix", outvcf, "-f"))
}
