==> caQTL_mapping_pbenaglio/notebooks/6.3_cicero_annotate.ipynb <==
suppressPackageStartupMessages(library(data.table))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library(stringr))
suppressPackageStartupMessages(library(RColorBrewer))

dir = '/nfs/lab/projects/pbmc_snATAC/analysis_v2/run_cicero_paola/'

setwd(dir)

# infile   = "final_conns/coaccessible.matrix"
# outfile  = "final_conns/coaccessible.bedpe"
# annofile = "final_conns/annotated.coaccessible.bedpe"
# B   = '/nfs/lab/publicdata/gencode_v19/gencode.v19.2kb_all_possible_transcripts.merged.bed'
# sed = paste("awk -F \"\\t\" '{print $1,$2}'", infile, "| tail -n +2 | sed 's/_/\\t/g' | pgltools sort >" , outfile) 
# pgl = paste('pgltools intersect1D -a', outfile ,'-b' , B,  '-wa -allA >', annofile)

#system(sed)
#system(pgl)

cells = c(  'b', "mono", 't',"nk")

anno       = read.table(annofile)
anno$loops = paste(anno[,1],anno[,2],anno[,3],anno[,4],anno[,5],anno[,6], sep="_")
anno       = anno[!duplicated(anno),] #remoeve two prom of the same gene
ag         = aggregate(V7~loops, anno, function(x) c("PP", "CP") [((sum( x=='A')/length(x)) %in% c(0,1)) +1] )
coac       = read.table("final_conns/coaccessible.matrix4s", header=T)
totc       = colSums(!is.na(coac[,cells]))
coac$loops = paste(coac$Peak1, coac$Peak2, sep="_")
coac       = merge(coac, ag, by="loops")
coac       = merge(coac, anno[,7:9], by="loops")
colnames(coac)[8:10] = c("type", "prom_side", "genes")

write.table(coac, "final_conns/annotated.coaccessible.matrix4s", sep="\t", quote=F, row.names=F)

stats =  data.frame()
for (ce in cells){
    
    mycel = coac[!is.na(coac[,ce]), c('loops', "type", ce)]
    mycel = mycel[!duplicated(mycel),]
    stats = rbind(stats, table(mycel$type)) 
}

colnames(stats) = c("CP", "PP")
rownames(stats) = cells

stats$tot = totc

stats

### filter for cell peaks No

#coacall       = read.table(infile, header=T)

# statsf =  data.frame()
# SUB    = data.frame()
# for (ce in cells){
# fbed = paste0('/nfs/lab/projects/pbmc_snATAC/analysis_v2/rasqual_eur/broad/', ce, '/' , ce, '.filtered.bed')

# fpeaks = read.table(fbed)
# fpeaks = paste(fpeaks[,1], fpeaks[,2]-1, fpeaks[,3], sep="_")
# mycel = coac[!is.na(coac[,ce]), c('loops', "type",'Peak1', 'Peak2', ce)]

# mycel =subset(mycel, Peak1 %in% fpeaks & Peak2 %in% fpeaks)
# mycel = mycel[!duplicated(mycel),c('loops', "type", ce)]

# sub     = subset(coacall, Peak1 %in% fpeaks & Peak2 %in% fpeaks )
# statsf  = rbind(statsf, c(table(mycel$type), sum(!is.na(sub[,ce]))))
    
# sub               = sub[, c('Peak1', 'Peak2', ce)]
# colnames(sub) [3] = 'coacc'
# sub$cell          = ce
# sub               = sub[complete.cases(sub),]
# SUB               = rbind(SUB, sub)
#     }

#colnames(statsf) = colnames(stats)
#rownames(statsf) = rownames(stats)

# widef  = spread(SUB, key=cell, value=coacc)
# widef  = as.data.frame(widef)
# write.table(widef, paste0('final_conns/peak_filtered.coaccessible.matrix'), sep= "\t", quote=F, row.names=F)
# coac = merge(widef, coac[,c('Peak1','Peak2','type','prom_side','genes')], by=c('Peak1','Peak2'))
# write.table(coac, paste0('final_conns/peak_filtered.annotated.coaccessible.matrix'), sep= "\t", quote=F, row.names=F)


pdf("../figures/SummaryCoaccessibleSites.pdf")
stat2  = round(stats/1000,1)  ## unfiltered
#stat3 = statsf/1000  ## filtered for coverage
par(xpd = TRUE, mfcol=c(3,3), las=2, mar = c(4,5,4,4))
bp<-barplot(stat2[,3],  ylab ="# coaccessible sites (x1e3)",names.arg = cells,ylim=c(0,4e+2))
text(bp, stat2[,3],labels=stat2[,3],pos=3)
bp<-barplot(stat2[,1],  ylab ="# coaccessible\n enhancer-promoters(x1e3)",names.arg = cells, ylim=c(0,120))
text(bp, stat2[,1],labels=stat2[,1],pos=3)
bp<-barplot(stat2[,2],  ylab ="# coaccessible\n promoter-promoters(x1e3)",names.arg = cells,ylim=c(0,40))
text(bp, stat2[,2],labels=stat2[,2],pos=3)
# barplot(stat3[,3],  ylab ="# coaccessible sites (x1e3)",names.arg = cells,ylim=c(0,4e+2))
# barplot(stat3[,1],  ylab ="# coaccessible\n enhancer-promoters(x1e3)",names.arg = cells, ylim=c(0,1.2e+2))
# barplot(stat3[,2],  ylab ="# coaccessible\n promoter-promoters(x1e3)",names.arg = cells, ylim = c(0,40))
dev.off()

### Annotate caQTLs

setwd('/nfs/lab/projects/pbmc_snATAC/analysis_v2/')

allqtls = read.table ("summarized_caqtls/run3/EUR_caqtl_leads_run3.tsv", header=T)
allqtls = subset(allqtls, flag_fdr10==TRUE & exclude==FALSE)

clrs = read.table("peaks/color_scheme.tsv")
clrs = rbind(clrs, matrix( c('orange2', 'b','tan4','nk','#7FC97F','mono', 'magenta4', "t"), ncol=2, byrow=T))

colcol  = data.frame(celltypes = clrs$V2, 
                     group= c('mono', 'mono', 'mono', "mono","b", "b", "nk", "nk",
                               't','t','t','t','t','t','mkc','pDC', 'plasma','b','nk','mono','t'))

allqtls$group = colcol$group[match (allqtls$cell, colcol$celltypes  )]

cells = c(  'b', "mono", 't',"nk")

## 1) Annotate all peaks
A = '/nfs/lab/publicdata/gencode_v19/gencode.v19.2kb_all_possible_transcripts.bed'
B = paste0('peaks/pbmc.sorted.merged.bare.bed')
C = paste0('peaks/pbmc.sorted.merged.prom.annot.bed')
#system(paste("bedtools intersect -a", A, "-b", B, "-wa -wb >", C))


pran =read.table(C)
pran$peak = paste0(pran[,5], ":",pran[,6], "-",pran[,7])
prang = aggregate(V4~peak, pran, function(b) paste(unique(b), collapse=","))

coac       = read.table("run_cicero_paola/final_conns/annotated.coaccessible.matrix4s", header=T)

cells

ptomoter_qtls  = data.frame()
summary_annot  = data.frame()
genes_enhancer =  data.frame()
for (c in cells) {

## 1) Select caQTL to test
results = subset(allqtls, group==c)
results$peak = gsub( ":|-",  "_", results$Feature)
results = results[!duplicated(results$Feature),]


## 2) Pair with the coaccessible peak (only one qtl)
cic   = coac[!is.na(coac[,c]), c('Peak1','Peak2', c)]

caca = cic[apply(cic[,1:2], 1, function(x)  sum(x %in% results$peak)==2 ),] #### complex coaccessible ca qtls
cic = cic[apply(cic[,1:2], 1, function(x)  sum(x %in% results$peak)==1 ),]
cic$Feature  = unlist(apply(cic[,1:2],  1, function(x)  x [x %in% results$peak]))
cic$coacPeak = unlist(apply(cic[,1:2], 1, function(x)  x [!(x %in% results$peak)]))
cic = cic[!duplicated(cic),]                         

fe = str_split_fixed(cic$Feature,"_",3)
co = str_split_fixed(cic$coacPeak,"_",3)
tab = data.frame(Feature = paste0(fe[,1],":" ,fe[,2],"-",fe[,3]), 
                 coacPeak = paste0(co[,1],":" ,co[,2],"-",co[,3]), coaccess = cic[,c], 
                 dist =abs(as.numeric(fe[,3]) - as.numeric(co[,3])) )


tab$feature_gene = prang$V4[match(tab$Feature, prang$peak)]
tab$coac_gene = prang$V4[match(tab$coacPeak, prang$peak)]
tab$type =c( "EP","P", "PP")  [rowSums(!is.na(tab[,c('feature_gene','coac_gene')])) + !(is.na(tab$feature_gene))]



##remove connsections to the same gene
tab = tab[!duplicated(tab[c("Feature", 'feature_gene','coac_gene','type')]),]

ag2        = aggregate(coac_gene~Feature, tab[tab$type=="EP",], function (x) paste(unique(x), collapse = ","))
ag2$gene_n = sapply( ag2$coac_gene, function(x) str_count(x, ',')+1)
ag2$gene_n[ag2$gene_n>5]<-"6+"
tb2 = table(ag2$gene_n)

## add promoter only
remains = subset(results, !(Feature %in% tab$Feature))
remains$feature_gene = prang$V4[match(remains$Feature, prang$peak)]
remains$type = "P"
remains = subset( remains, !is.na(remains$feature_gene))
tab     = merge(tab, remains[,c('Feature', 'feature_gene', "type")], by=c('Feature', 'feature_gene', "type"), all=T)
tab$group=c

ag = aggregate(type~ Feature,tab, function(b) paste(unique(b), collapse=","))
ag$type[grepl("PP", ag$type )]<-"PP"
tb1 = table(ag$type)
tb1["CC"]<- sum(!duplicated(caca$Peak1, caca$Peak2))  ### complex cases


tb1["NP"] <- sum(!(results$Feature %in% tab$Feature)) - sum(!duplicated(caca$Peak1, caca$Peak2)) ## not promoter
  
               
ptomoter_qtls  = rbind(ptomoter_qtls,tab)
summary_annot  = rbind(summary_annot, tb1)
genes_enhancer = rbind(genes_enhancer, tb2)
               }

colnames(summary_annot)  = paste(c(2,1,3,4,5),names(tb1), sep="_")
colnames(genes_enhancer) = names(tb2)
rownames(summary_annot)  = rownames(genes_enhancer) = cells

summary_annot

genes_enhancer

genes_enhancer[,1]/rowSums(genes_enhancer)

write.table(ptomoter_qtls, "summarized_caqtls/Promoter_associated_caQTLs.tsv", sep="\t", row.names=F, quote=F)

pal = pal = brewer.pal(6,'RdYlBu')[c(1:3,5)]

toplot = t(summary_annot[, c(2,1,3,4)])

toplot

toplot[2,]/colSums(toplot[1:3,])

pdf("figures/caQTL_prom_annotat.pdf")
#layout(matrix(c(1,1,1,1,1,2,2,2,2,2,3,4,5,6,7,8,9,10,11,12), ncol=2, byrow = F))
layout(matrix(c(1,1,1,1,1,    2,3,4,5,  6,6,6,7,7,7,8,8,8      ), ncol=2, byrow = F))
par( las=2, mar = c(4,4,2,8))
barplot(toplot, ylim=c(0,2500), ylab="# caQTL-peaks", col=pal)
legend("topleft", legend = rownames(toplot), col=pal, pch=15, bty=F, pt.cex=2)
#plot.new()
par( las=1, mar = c(2,8,0,8))

for( n in 1:4){  
    barplot(as.matrix(genes_enhancer[n,]), ylim = c(0, max(genes_enhancer[n,]) + (max(genes_enhancer[n,]) /7 ) ) )
    mtext(rownames(genes_enhancer[n,]),at = 1, line=-0.5)
}

# ### Add stats prom before
# par(xpd = TRUE, las=2, mar = c(4,8,4,8))
# bp<-barplot(stat2[,3],  ylab ="# coaccessible sites (x1e3)",names.arg = cells,ylim=c(0,4e+2))
# text(bp, stat2[,3],labels=stat2[,3],pos=3)
# bp<-barplot(stat2[,1],  ylab ="# coaccessible\n enhancer-promoters(x1e3)",names.arg = cells, ylim=c(0,120))
# text(bp, stat2[,1],labels=stat2[,1],pos=3)
# bp<-barplot(stat2[,2],  ylab ="# coaccessible\n promoter-promoters(x1e3)",names.arg = cells,ylim=c(0,40))
# text(bp, stat2[,2],labels=stat2[,2],pos=3)
 dev.off()

ptomoter_qtls = read.table("summarized_caqtls/Promoter_associated_caQTLs.tsv",header=T)

sum(!duplicated(ptomoter_qtls$Feature))

head(ptomoter_qtls)

ppp = subset(ptomoter_qtls, type!="EP")

sum(!duplicated(ppp$Feature))

1281/6901

coc = read.table("run_cicero_paola/final_conns/annotated.coaccessible.matrix4s", header=T)

sum(!duplicated(coc$loops[coc$type=='PP']))

sum(!duplicated(coc$loops[coc$type=='CP']))
