==> caQTL_mapping_pbenaglio/notebooks/6.4_Cicero_plot_examples.ipynb <==
suppressPackageStartupMessages(library(cicero))
suppressPackageStartupMessages(library(stringr))
suppressPackageStartupMessages(library(data.table))
suppressPackageStartupMessages(library(parallel))

setwd('/nfs/lab/projects/pbmc_snATAC/analysis_v2/run_cicero_paola')

gene_annotation <- read.table('gencode.v19.cicero_gene_annotation.txt', header=T, stringsAsFactors=FALSE)
gene_annotation <- gene_annotation[gene_annotation[['feature']]=='protein_coding',]

filtmat = read.table("final_conns/annotated.coaccessible.matrix4s", header=T)

cp = subset(filtmat,type=="CP")

connst = cp[,c('Peak1', 'Peak2', 't')]
#connst = filtmat[,c('Peak1', 'Peak2', 't')]
colnames(connst)[3]= "coaccess"
connst = connst[complete.cases(connst),]

pdf("Example_GSC.pdf")
plot_connections(connst, 'chr14', 95216075 ,95432916, 
                 connection_color = "red",
                 viewpoint =  'chr14:95411884-95413043',
                gene_model = gene_annotation, 
                coaccess_cutoff = 0.05, alpha_by_coaccess = TRUE,
                connection_width = 2, 
                collapseTranscripts = "longest" )
dev.off()

chr6:90,038,953-91,077,831

pdf("Example_BACH2.pdf")
plot_connections(connst, "chr6", 90700000,91200000,
                 connection_color = "red",
                 viewpoint =  'chr6_90976592_90977416',
                gene_model = gene_annotation, 
                coaccess_cutoff = 0.05, alpha_by_coaccess = TRUE,
                connection_width = 2, 
                collapseTranscripts = "longest" )
dev.off()

pdf("Example_CXCR5.pdf")
plot_connections(connst, "chr11", 118400000,118900000,
                 connection_color = "red",
                 viewpoint =  'chr11_118743054_118743481',
                gene_model = gene_annotation, 
                coaccess_cutoff = 0.05, alpha_by_coaccess = TRUE,
                connection_width = 2, 
                collapseTranscripts = "longest" )
dev.off()

connsb = cp[,c('Peak1', 'Peak2', 'b')]
colnames(connsb)[3]= "coaccess"
connsb = connsb[complete.cases(connsb),]

connsm = cp[,c('Peak1', 'Peak2', 'mono')]
colnames(connsm)[3]= "coaccess"
connsm = connsm[complete.cases(connsm),]

connsn = cp[,c('Peak1', 'Peak2', 'nk')]
colnames(connsn)[3]= "coaccess"
connsn = connsn[complete.cases(connsn),]

pdf("Example_NINJ2.pdf")
plot_connections(connst, "chr12", 295800,1138400, comparison_connection_color = "red",
                 connection_color = "red",
                 viewpoint =  'chr12_739181_740753',
                 comparison_track = connsm,
                gene_model = gene_annotation, 
                coaccess_cutoff = 0.05, alpha_by_coaccess = TRUE,
                connection_width = 2, 
                collapseTranscripts = "longest" )
dev.off()

pdf("Example_cc20.pdf")
plot_connections(connst, "chr2", 227830402,229027101, #comparison_connection_color = "red",
                 connection_color = "red",
                 viewpoint =  'chr2_228669766_228670866',
                # comparison_track = connsm,
                gene_model = gene_annotation, 
                coaccess_cutoff = 0.05, alpha_by_coaccess = TRUE,
                connection_width = 2, 
                collapseTranscripts = "longest" )
dev.off()
