==> caQTL_mapping_pbenaglio/notebooks/7_Fine_mapping_plots.ipynb <==
library(stringr)
library(scales)
library(RColorBrewer)

### Plot credible sets

traits = c( 'Asthma_child_onset.99credset.PPA.bed',
            'Atopic_dermatitis.99credset.PPA.bed',
            'Autoimmune_vitiligo.99credset.PPA.bed',
            'Basophil_count.99credset.PPA.bed',
            'Crohns_disease.99credset.PPA.bed',
            'Eosinophil_count.99credset.PPA.bed',
            'Gout.99credset.PPA.bed',
            'Lymphocyte_count.99credset.PPA.bed',
            'Monocyte_count.99credset.PPA.bed',
            'Neutrophil_count.99credset.PPA.bed',
            'Primary_sclerosing_cholangitis.99credset.PPA.bed',
            'Rheumatoid_arthritis.99credset.PPA.bed',
            'Selective_IgA_deficiency.99credset.PPA.bed',
            'Systemic_lupus_erythematosus.99credset.PPA.bed',
            'Type_1_diabetes.99credset.PPA.bed',
            'Ulcerative_colitis.99credset.PPA.bed')

fmdir = '/nfs/lab/projects/pbmc_snATAC/data/credible_sets/'

figdir = "/nfs/lab/projects/pbmc_snATAC/analysis_v2/figures/"

pl = data.frame()

df = data.frame()
for (i in 1: length(traits)){
fm = read.table(paste0(fmdir, traits[i] ))
ag = aggregate(V4~V5, fm, length)
ag$br = with(ag, cut(V4, breaks = c(1,3,6,21,51,101, max(V4)+1), right=F))
pl = rbind(pl, table(ag$br))
    ag$trait = i
    df = rbind(df,ag )
    }

colnames(pl) = names(table(ag$br))
rownames(pl) = str_split_fixed(traits, "\\.", 2)[,1]
totals = rowSums(pl)

pct = pl/totals
pal = rev(brewer.pal(6, "Purples"))

mat =as.matrix(t(pl[order(totals),]))
tots = totals [order(totals)]
leg = c("1-2", "3-5","6-20", "21-50" , "51-100", "101+")

pdf(paste0(figdir,"Barplot_loci_smaller.pdf"))
par(mar=c(8,12,10,6), xpd = TRUE)
bp = barplot(mat, las=1, horiz = T, col=pal, xlim=c(0,200), xlab="# loci")
text(tots,bp, labels = tots, pos = 4)
legend("bottomright",legend = leg, pch=22, pt.cex = 1.5,
       pt.bg = pal, title="# variants per locus")
dev.off()

sum(totals)

summary(df[,2])

ss = split(df, df$trait)
sapply (ss, function(x) summary(x[,2]))

### caQTLs in credible sets
include also variants with the same q-value as the lead-

alltraits = data.frame()
for (i in 1: length(totals)){

fm = read.table(paste0(fmdir, traits[i] ))[,4:7]
colnames(fm) = c("varID", "Locus", "LNBF", "PPA")
 fm$Trait = paste(sprintf("%02d",rank(totals,ties.method ='first')[i]), str_split_fixed(traits[i], "\\.", 2)[,1], sep="_")
 alltraits = rbind(alltraits, fm)
    }
    

caqdir = '/nfs/lab/projects/pbmc_snATAC/analysis_v2/summarized_caqtls/run3/'
files  = list.files(caqdir)[grepl ('_caqtl_sumstats_run3.tsv', list.files(caqdir))] 
files = files[!(files %in% c('cd4_t_caqtl_sumstats_run3.tsv', 'cd8_t_caqtl_sumstats_run3.tsv', "bulk_caqtl_sumstats_run3.tsv"))]
cells  = gsub( '_caqtl_sumstats_run3.tsv', "" , files)

colkeep = c('Feature','varID', 'rsID', 'P_value', 'Q_value' ,'Effect_size' ,'cell' )

M= data.frame()
for (y in 1:length(files)){

sms = read.table(paste0(caqdir, files[y]), header=T)
sss = subset(sms, Is_peak_caQTL==T & exclude==F)
sss = sss[order(sss$Q_value),]
ag  = aggregate(Q_value~Feature, sss, min)
mm  = merge (sss, ag, by= colnames(ag))
mm$cell = cells[y]
M    = rbind(M, mm[,colkeep])
}

subset(M, rsID=="rs34038797")

allcaq = subset(alltraits,varID %in% c(substring(M$varID,4), as.character(M$rsID)) ) 
## this is only bc the t1d from tony had rsids

agc = aggregate(varID~Trait,allcaq, length )

pdf(paste0(figdir, "Finemapped_caQTLs_bxpl_newres.pdf"))
par(mar=c(8,6,10,0), xpd = TRUE, mfrow=c(1,2))
#stripchart(PPA~Trait,allcaq , pch=19, las=1, col=alpha(pal[1],0.5))
boxplot(PPA~Trait,allcaq , horizontal = T, las=1, col=alpha(pal[1],0.5), pch=16,
        outcol= alpha(pal[1],0.5))
par(mar=c(8,6,10,0))
barplot(agc[,2], horiz=T, names.arg =agc[,2] , las=1)
dev.off()

allcaq[allcaq$Locus=="BACH2",]

sum(!duplicated(allcaq$varID))

sum(!duplicated(allcaq$Locus))

### Annotate with promoters and coaccessible promoters

setwd("/nfs/lab/projects/pbmc_snATAC/analysis_v2/")

map = read.table("summarized_caqtls/Promoter_associated_caQTLs.tsv", header=T)

clrs = read.table("peaks/color_scheme.tsv")
clrs = rbind(clrs, matrix( c('orange2', 'b','tan4','nk','#7FC97F','mono', 'magenta4', "t"), ncol=2, byrow=T))
colcol  = data.frame(celltypes = clrs$V2, 
                     group= c('mono', 'mono', 'mono', "mono","b", "b", "nk", "nk",
                               't','t','t','t','t','t','mkc','pDC', 'plasma','b','nk','mono','t'))

M$group = colcol$group[match (M$cell, colcol$celltypes  )]

groups = c( 'b', "mono", 't',"nk")  ### mkc are few and do not overlap any finemapped variant

#sum(c(unique(subset(M, cell == "mkc")$varID), unique(subset(M, cell == "mkc")$rsID)) %in% allcaq$varID)

annoq = data.frame()
for (gr in groups){

test = subset(map, group==gr)
    
agr  = aggregate(type~feature_gene, test[test$type!="EP",], function(x)  c('P','PP')[max(str_count(x))]  ) ## this is to remove extra P annot
test2 = merge(test[test$type!="EP",], agr, by=colnames(agr))
test3 = rbind(test2, test[test$type=="EP",])
                 
qt = subset(M, group==gr)
qt = merge(qt, test3, by=c('Feature','group'))
annoq = rbind(annoq,qt)                 
                 }
                 

annoq$varID= substring(annoq$varID,4)

allcaq = subset(allcaq, PPA >0.01)

mm = merge(allcaq, annoq, by="varID")
mmt1d = merge( annoq, allcaq, by.y="varID", by.x="rsID")

mm = rbind (mm, mmt1d)
mm = mm[!duplicated(mm),]

#subset(mm, rsID=='rs72928038')

#subset(mm, rsID=='rs12365699')

#subset(mm, rsID=='rs7731626')

#subset(mm, rsID=='rs17293632')

#subset(mm,coac_gene=="SMAD3" )

#subset(mm, rsID=='rs34038797')

rmdu =mm[!duplicated(mm[,c("varID", "Trait", "type")]),]

dim(rmdu)

#rmdu =mm[!duplicated(mm[,c("varID", "Trait")]),]

#dim(rmdu)

TA = table(rmdu$Trait, rmdu$type)

extra = unique(alltraits$Trait)[!(unique(alltraits$Trait) %in% rownames(TA))]
TA = rbind(rbind(TA, 0),0)
rownames(TA)[15:16]=extra
TA = TA[order(rownames(TA)),]

colnames(TA) = c('2_EP','1_P','3_PP')
TA = TA[,order(colnames(TA))]

pal2 = brewer.pal(6,'RdYlBu')[c(1:3,5)]
pdf(paste0(figdir, "CAtegory_caQTLs_finemap_newresults.pdf"))
par(mar=c(8,6,10,2), xpd = TRUE, mfrow=c(1,2))
plot.new()
barplot(t(TA), horiz = T, las=1, col=pal2, xlim=c(0,35))
legend("bottomright",legend = colnames(TA), pch=22, pt.cex = 1.8,
       pt.bg = pal2, title="#caQTLs\n(PPA>0.01)", bty="n")
dev.off()

### GWAS plots

setwd('/nfs/lab/projects/pbmc_snATAC/analysis_v2/gwas_sumstats')

# loci
# chr6   89207850   92324486
# chr11   117991786   119494785
# chr2 227431691 229908940

# commands:
# - bedtools intersect -a /nfs/lab/publicdata/1kg_snps/ALL.wgs.phase3_sites.bed.ids -b loci -wa > loci_snps.bed
# - awk -v OFS="\t" '{print $4}' loci_snps.bed > loci_snps.rsid_list
# - zgrep -F -f loci_snps.rsid_list T1D.Aylward_2018_biorxiv.txt.gz > T1D.Aylward_2018.select

t1d  = read.table("T1D.Aylward_2018.select", sep="\t")
ly   = read.table("Astle.Cell.2016.lymph.select", sep="\t")
as   = read.table("Ferreira.2019.AJHG.Child_onset_asthma.select")

uc   = read.table("deLange.NatGenet.2017.Ulcerative_Colitis.select")

coo  = read.table("lociAll_snps.bed", sep="\t")
colnames(coo) = c( "chr", "start", "pos", 'rsid', 'id')

t1d = merge(t1d, coo, by.x=1, by.y=4)
t1d = subset(t1d, chr == 6)
as = merge(as, coo, by.x=1, by.y=4)
as = subset(as, chr == 11)
ly = merge(ly, coo, by.x=1, by.y=4)
ly = subset(ly, chr == 6)

uc = merge(uc, coo, by.x=1, by.y=4)
uc = subset(uc, chr == 2)

uc$log10pval = -log10(uc$V9)
uc$credset = uc$id %in% alltraits$varID[alltraits$Trait == "10_Ulcerative_colitis" & alltraits$Locus=="2:228670476:C:G" ]

t1d$log10pval = -log10(t1d$V9)
t1d$credset = t1d$V1 %in% alltraits$varID[alltraits$Trait == "08_Type_1_diabetes" & alltraits$Locus=="BACH2" ]

as$log10pval = -log10(as$V13)
as$credset = as$id %in% alltraits$varID[alltraits$Trait == "13_Asthma_child_onset" & alltraits$Locus=='11:118743286:G:A']

ly$log10pval = -log10(ly$V9)
ly$credset = ly$id %in% alltraits$varID[alltraits$Trait == "14_Lymphocyte_count"& alltraits$Locus=='6:90976768:G:A']

t1d = t1d[order(-t1d$log10pval),]
as = as[order(-as$log10pval),]
ly = ly[order(-ly$log10pval),]
uc = uc[order(-uc$log10pval),]

t1d$caQTL = t1d$id %in% allcaq$varID
as$caQTL = as$id %in% allcaq$varID
ly$caQTL = ly$id %in% allcaq$varID
uc$caQTL = uc$id %in% allcaq$varID

pdf(paste0(figdir, "GWAS_loci_mainfigure.pdf"))
par(mfrow=c(2,2), mar = c(4,4,10,0), las=1)
plot(log10pval~V3, t1d, pch=c(19,17)[t1d$caQTL +1], col=alpha(c("gray", "red"),0.5)[t1d$credset +1], cex=0.6,
    xlim=c(90700000,91200000), main="BACH2 - T1D")

plot(log10pval~V3, ly, pch=c(19,17)[ly$caQTL +1], col=alpha(c("gray", "red"),0.5)[ly$credset +1], cex=0.6,
    xlim=c(90700000,91200000), main="BACH2 - Lymphocyte count")

plot(log10pval~V3, as, pch=c(19,17)[as$caQTL +1], col=alpha(c("gray", "red"),0.5)[as$credset +1], cex=0.6,
  xlim=c(118400000,118900000), main="CXCR5 locus- Asthma")

plot(log10pval~V3, uc, pch=c(19,17)[uc$caQTL +1], col=alpha(c("gray", "red"),0.5)[uc$credset +1], cex=0.6,
     #xlim=c(228563194, 228887101), main="CCL20-Ulcerative Colitis")
     xlim=c(227830402,229027101))
  #xlim=c(228001852,228793788), main="CCL20-Ulcerative Colitis")
dev.off()

png(paste0(figdir, "GWAS_bach2.png"), width = 4, height = 1.6, units = 'in', res = 300)
par( pin=c(4,1.6), las=1)
plot(log10pval~V3, ly, pch=c(19,17)[ly$caQTL +1], col=alpha(c("gray", "red"),0.5)[ly$credset +1], cex=0.6,axes=F,
    xlim=c(90700000,91200000))
dev.off()

png(paste0(figdir, "GWAS_CXCR5.png"), width = 4, height = 1.6, units = 'in', res = 300)
par( pin=c(4,1.6), las=1)
plot(log10pval~V3, as, pch=c(19,17)[as$caQTL +1], col=alpha(c("gray", "red"),0.5)[as$credset +1], cex=0.6,axes=F,
  xlim=c(118400000,118900000))
dev.off()

png(paste0(figdir, "GWAS_CCL20.png"), width = 4, height = 1.6, units = 'in', res = 300)
par( pin=c(4,1.6), las=1)
plot(log10pval~V3, uc, pch=c(19,17)[uc$caQTL +1], col=alpha(c("gray", "red"),0.5)[uc$credset +1], cex=0.6,axes=F,
  xlim=c(227830402,229027101))
dev.off()

### Supptables

getwd()

st0 = read.table("summarized_caqtls/run3/EUR_caqtl_leads_run3.tsv", header=T)

dim(st0)

st = subset(st0, flag_fdr10==T)

dim(st)

st = st[order( st$cell,st$P_VAl),]

write.csv(st,"summarized_caqtls/run3/EUR_caqtl_leads_subset.csv", row.names=F)

head(st)

sum(st$exclude)/ nrow(st)

st$sig = st$flag_fdr10 & st$exclude==FALSE

ag1 = aggregate(sig~cell, st, sum)

ag2=aggregate(Feature~cell, st0, length)

agg = merge (ag1, ag2, by=1)

agg$fract = agg$sig/agg$Feature

agg[order(agg$sig),]
