==> caQTL_mapping_pbenaglio/notebooks/7_Finrich_enrichment_credible_sets.ipynb <==
library(stringr)
library(parallel)
library(tidyr)

traits = c( 'Asthma_child_onset.99credset.PPA.bed',
            'Atopic_dermatitis.99credset.PPA.bed',
            'Autoimmune_vitiligo.99credset.PPA.bed',
            'Basophil_count.99credset.PPA.bed',
            'Crohns_disease.99credset.PPA.bed',
            'Eosinophil_count.99credset.PPA.bed',
            'Gout.99credset.PPA.bed',
            'Lymphocyte_count.99credset.PPA.bed',
            'Monocyte_count.99credset.PPA.bed',
            'Neutrophil_count.99credset.PPA.bed',
            'Primary_sclerosing_cholangitis.99credset.PPA.bed',
            'Rheumatoid_arthritis.99credset.PPA.bed',
            'Selective_IgA_deficiency.99credset.PPA.bed',
            'Systemic_lupus_erythematosus.99credset.PPA.bed',
            'Type_1_diabetes.99credset.PPA.bed',
            'Ulcerative_colitis.99credset.PPA.bed')

fmdir = '/nfs/lab/projects/pbmc_snATAC/data/credible_sets/'

out = '/nfs/lab/projects/pbmc_snATAC/analysis_v2/finrich/99credset_immune/'

setwd('/nfs/lab/projects/pbmc_snATAC/analysis_v2')

results = data.table::fread("summarized_caqtls/run3/EUR_caqtl_leads_run3.tsv", header=T)

celltypes = unique(results$cell)[unique(results$cell)!='bulk']

outdir = '/nfs/lab/projects/pbmc_snATAC/analysis_v2/finrich/bedfiles/'

for (c in celltypes) {
cell_results      = subset(results, cell == c)
step1 = str_split_fixed(cell_results$Feature, "\\:", 2 )
bed   = cbind(step1[,1], str_split_fixed(step1[,2], "\\-", 2 ))

write.table(bed[cell_results$flag_fdr10==TRUE & cell_results$exclude==FALSE,], paste0(outdir, c, ".caqtl_peaks.bed"), sep="\t", quote=F, col.names=F, row.names=F)
write.table(bed[cell_results$flag_fdr10==FALSE,], paste0(outdir, c, ".non_caqtl_peaks.bed"), sep="\t", quote=F, col.names=F, row.names=F)

    }


beds_pos = paste0(outdir, list.files(outdir)[!(grepl(list.files(outdir),pattern = ".non_caqtl_peaks.bed"))])
beds_neg = paste0(outdir, list.files(outdir)[grepl(list.files(outdir),pattern = ".non_caqtl_peaks.bed")])

resdir   = '/nfs/lab/projects/pbmc_snATAC/analysis_v2/finrich/results/'

outfiles = paste0(resdir, 
                  str_split_fixed(list.files(outdir)[grepl(list.files(outdir),pattern = ".non_caqtl_peaks.bed")], "\\.", 2)[,1])

for (tr in list.files(fmdir)){

 mclapply(1:length(outfiles), function(x) system(paste0(
    "finrich ", fmdir, tr, " ", beds_pos[x]," " ,beds_neg[x], 
        " --permutations 10000 --processes 8 > ", outfiles[x], ".", tr, ".json" )) ,mc.cores=32)
}

#### Read in the results

library('rjson')

infiles = list.files(resdir)[grepl("PPA", list.files(resdir))]

m = matrix(NA, nrow = length(infiles), ncol=6)
for( rr in 1:length(infiles)) {
    file= infiles[rr]
    test = readLines(paste0(resdir,file))
    if(length(test)==0){
        m[rr,1:4] = NA} else {
        
    if ( grepl('Infinity' , test)== TRUE   ){
    m[rr,1:4] = c(1,0,0,0)  
    }else{
    js = fromJSON(file =paste0(resdir,file))
    m[rr,1:4] = unlist(js) [c("pval", "logOR", "conf_lower", "conf_upper")]
    }
        }
    m[rr,5:6] = str_split_fixed(file, "\\.", 6)[,1:2]
}

colnames(m) = c("pval", "logOR", "conf_lower", "conf_upper", "cell", "trait")

m2 = data.frame(m)

macro = as.character(unique(m2$cell[m2$cell %in% c("t", "b", "nk", "mono")]))
micro = as.character(unique(m2$cell[!(m2$cell %in% c("t", "b", "nk", "mono"))]))

m2 = m2[complete.cases(m2),]

wideor = spread(m2[,c('trait','cell','logOR' )],  cell, logOR)
widepv = spread(m2[,c('trait','cell','pval' )],  cell, pval)
wideor[is.na(wideor)] <-0
widepv[is.na(widepv)] <-1
or = apply(wideor[,-1], 2, as.numeric)
pv = apply(widepv[,-1], 2, as.numeric)

min(pv[pv>0])

pv[pv==0]<-min(pv[pv>0])

rownames(pv) = rownames(or) = widepv$trait

signedpv = -log10(pv)
signedpv = signedpv*sign(or)

library(pheatmap)

ix = pv
ix[pv<0.05]<-"*"
ix[pv<0.005]<-"**"
ix[pv<0.0005]<-"X"
ix[pv>=0.05]<-""

traits = c( 'Monocyte_count','Eosinophil_count','Lymphocyte_count','Asthma_child_onset',
            'Crohns_disease', 'Neutrophil_count','Ulcerative_colitis','Rheumatoid_arthritis',
             "Type_1_diabetes",'Autoimmune_vitiligo','Basophil_count','Systemic_lupus_erythematosus',
            'Gout','Atopic_dermatitis', 'Primary_sclerosing_cholangitis','Selective_IgA_deficiency')


cat(paste(micro, collapse = "\', \'"))

micro =c('naive_cd4_t', 'act_cd4_t', 'naive_cd8_t', 'cyto_cd8_t','mem_cd8_t','tReg',
  'adaptive_NK','cyto_nk',  'cMono', 'ncMono', 
   'iMono', 'cDC', 'mem_b',   'naive_b', 'mkc')


library(RColorBrewer)

col = rev(colorRampPalette(brewer.pal(9, "RdBu"))(50))
col = rev(colorRampPalette(brewer.pal(9, "PuOr"))(50))

mat = as.matrix(or[traits,micro])
pheatmap(mat,    show_colnames = TRUE, angle_col = 45, fontsize_col = 8,
         show_rownames=TRUE, cluster_rows = F, cluster_cols = F,
         col = col,  breaks=seq(-abs(max(mat)), abs(max(mat)), length.out=50) ,
         cellheight = 8, cellwidth = 10, fontsize_row = 8, display_numbers= ix[traits,micro],
          filename =  'figures/Heatmap_OR_subtypes.pdf'    
        )

mat=as.matrix(signedpv[traits,micro])
pheatmap(mat,    show_colnames = TRUE, angle_col = 45,fontsize_col = 8,
         show_rownames=TRUE, cluster_rows = F, cluster_cols = F,
         col = col,  breaks=seq(-abs(max(mat)), abs(max(mat)), length.out=50) ,
         cellheight = 8, cellwidth = 10, fontsize_row = 8, display_numbers= ix[traits,micro],
          filename =  'figures/Heatmap_PV_subtypes.pdf',
        )
mat = as.matrix(or[traits,macro])
pheatmap(mat,    show_colnames = TRUE, angle_col = 45,fontsize_col = 8,
         show_rownames=TRUE, cluster_rows = F, cluster_cols = F,
         col = col,  breaks=seq(-abs(max(mat)), abs(max(mat)), length.out=50) ,
         cellheight = 8, cellwidth = 10, fontsize_row = 8, display_numbers= ix[traits,macro],
          filename =  'figures/Heatmap_OR.pdf'    
        )

mat=as.matrix(signedpv[traits,macro])
pheatmap(mat,    show_colnames = TRUE, angle_col = 45,fontsize_col = 8,
         show_rownames=TRUE, cluster_rows = F, cluster_cols = F,
         col = col,  breaks=seq(-abs(max(mat)), abs(max(mat)), length.out=50) ,
         cellheight = 8, cellwidth = 10, fontsize_row = 8, display_numbers= ix[traits,macro],
          filename =  'figures/Heatmap_PV.pdf',
        )


as.matrix(or[traits,micro])

as.matrix(pv[traits,micro])
